Create ensemble.py
